project_weather/dash/app/requirements.txt
dash_html_components
dash_core_components
dash
dash_bootstrap_components
plotly
dash_table
google.cloud